+++ packages/libs/eda/src/lib/core/components/computations/plugins/betadiv.tsx
@@ -21,7 +21,7 @@ import ScatterBetadivSVG from '../../visualizations/implementations/selectorIcon
 import { ComputationStepContainer } from '../ComputationStepContainer';
 import './Plugins.scss';
 import { makeClassNameHelper } from '@veupathdb/wdk-client/lib/Utils/ComponentUtils';
-import { VariableCollectionSelectList } from '../../variableSelectors/VariableCollectionSelectList';
+import { VariableCollectionSelectList } from '../../variableSelectors/VariableCollectionSingleSelect';
 
 const cx = makeClassNameHelper('AppStepConfigurationContainer');
