+++ packages/libs/eda/src/lib/core/components/variableSelectors/VariableCollectionSingleSelect.tsx
